ToDotNot unit test conversion - part 8.
